fix(country-code-test): vibrate on failure like URL test

+++ V2rayNG/app/src/main/java/com/v2ray/ang/handler/SettingsManager.kt
@@ -364,7 +364,7 @@ object SettingsManager {
 
     fun getCountryCodeTestTimeout(): Int {
         val seconds = MmkvManager.decodeSettingsString(AppConfig.PREF_COUNTRY_CODE_TIMEOUT)?.toIntOrNull() ?: 5
-        return (seconds.coerceIn(1, 15) * 1000).toInt()
+        return seconds.coerceIn(1, 15) * 1000        
     }
 
     fun getLocale(): Locale {
